Refactor code structure
Refactored the codebase for improved organization and maintainability. Consider reviewing the length of specific files for potential further refactoring.
[skip gpt_engineer]
