+++ src/components/home/UserMenu.tsx
@@ -1,0 +1,49 @@
+import { Settings, LogOut, MessageSquare } from "lucide-react"
+import { useNavigate } from "react-router-dom"
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger,
+} from "@/components/ui/dropdown-menu"
+import { supabase } from "@/integrations/supabase/client"
+
+export const UserMenu = () => {
+  const navigate = useNavigate()
+
+  const handleSignOut = async () => {
+    await supabase.auth.signOut()
+    navigate('/login')
+  }
+
+  return (
+    <div className="w-full flex justify-end mb-8">
+      <DropdownMenu>
+        <DropdownMenuTrigger asChild>
+          <div className="w-10 h-10 rounded-full bg-gray-200 flex items-center justify-center cursor-pointer hover:bg-gray-300 transition-colors">
+            <svg className="w-6 h-6 text-gray-600" viewBox="0 0 24 24">
+              <path 
+                fill="currentColor" 
+                d="M12 12c2.21 0 4-1.79 4-4s-1.79-4-4-4-4 1.79-4 4 1.79 4 4 4zm0 2c-2.67 0-8 1.34-8 4v2h16v-2c0-2.66-5.33-4-8-4z"
+              />
+            </svg>
+          </div>
+        </DropdownMenuTrigger>
+        <DropdownMenuContent align="end" className="w-48">
+          <DropdownMenuItem onClick={() => navigate('/suggestions')} className="cursor-pointer">
+            <MessageSquare className="mr-2 h-4 w-4" />
+            <span>Suggérer des fonctionnalités</span>
+          </DropdownMenuItem>
+          <DropdownMenuItem onClick={() => navigate('/settings')} className="cursor-pointer">
+            <Settings className="mr-2 h-4 w-4" />
+            <span>Paramètres</span>
+          </DropdownMenuItem>
+          <DropdownMenuItem onClick={handleSignOut} className="cursor-pointer text-red-600">
+            <LogOut className="mr-2 h-4 w-4" />
+            <span>Se déconnecter</span>
+          </DropdownMenuItem>
+        </DropdownMenuContent>
+      </DropdownMenu>
+    </div>
+  )
+}
